requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.12.1
colorama==0.4.6
compress-pickle==2.1.0
cryptography==43.0.1
Cython==3.0.11
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
gensim==4.3.3
google-auth==2.35.0
gunicorn==23.0.0
icecream==2.1.3
idna==3.10
itsdangerous==2.2.0
jellyfish==1.1.0
Jinja2==3.1.5
joblib==1.4.2
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pg8000==1.31.2
pgvector==0.3.4
pillow==10.4.0
pipe==2.2
pry.py==0.1.1
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxDamerauLevenshtein==1.8.0
regex==2024.9.11
requests==2.32.3
rsa==4.9
schedule==1.2.2
scipy==1.13.1
scramp==1.4.5
six==1.16.0
smart-open==7.0.4
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.13.1
openai==1.61.1
google-cloud-secret-manager==2.21.1
rapidfuzz==3.12.1
huggingface-hub==0.23.5
backoff==2.2.1
tenacity==9.0.0
sqlite-vec==0.1.6
